+++ packages/examples/mongodb-express/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @better-conversation/example-mongodb-express
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @better-conversation/core@1.0.1
+  - @better-conversation/adapter-mongodb@1.0.1
+  - @better-conversation/handler-express@1.0.1
+
 ## 0.0.1
 
 ### Patch Changes