+++ packages/examples/prisma-nextjs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @better-conversation/example-prisma-nextjs
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @better-conversation/core@1.0.1
+  - @better-conversation/adapter-prisma@1.0.1
+  - @better-conversation/handler-next@1.0.1
+
 ## 0.0.1
 
 ### Patch Changes